+++ java/benchmark/src/main/java/org/rocksdb/benchmark/DbBenchmark.java
@@ -575,11 +575,11 @@ public class DbBenchmark {
     options.setTargetFileSizeBase(
         (Integer)flags_.get(Flag.target_file_size_base));
     options.setTargetFileSizeMultiplier(
-        (Integer)flags_.get(Flag.target_file_size_multiplier));
+        (Double)flags_.get(Flag.target_file_size_multiplier));
     options.setMaxBytesForLevelBase(
         (Integer)flags_.get(Flag.max_bytes_for_level_base));
     options.setMaxBytesForLevelMultiplier(
-        (Integer)flags_.get(Flag.max_bytes_for_level_multiplier));
+        (Double)flags_.get(Flag.max_bytes_for_level_multiplier));
     options.setLevelZeroStopWritesTrigger(
         (Integer)flags_.get(Flag.level0_stop_writes_trigger));
     options.setLevelZeroSlowdownWritesTrigger(
@@ -1268,7 +1268,7 @@ public class DbBenchmark {
     max_bytes_for_level_multiplier(10,
         "A multiplier to compute max bytes for level-N (N >= 2)") {
       @Override public Object parseValue(String value) {
-        return Integer.parseInt(value);
+        return Double.parseDouble(value);
       }
     },
     level0_stop_writes_trigger(12,"Number of files in level-0\n" +
